package.json --> added @ngrx/effects package.
app.module.ts --> imported PictresqueEffects and Effects module.

modelcontent.ts --> this component dispatches an action to the reducer instead of directly
calling the pictresqueservices upload post method.

posts.component.ts --> gets posts from the store/state object instead of calling the pictresque services getposts method.

pictresque.actions.ts --> this is where all the actions reside.
actions are dispatched within any component that wants to call it.

pictresque.effects.ts --> this is where all the effects reside. effects are essentially
what are called by the action and call the services which pass the data from the result of the http request into the reducer which mutates the state.

pictresque.reducer.ts --> this mutates the state by creating a copy of the state and modifying it.